+++ src/components/Header/Header.test.js
@@ -5,7 +5,7 @@ import Header from './Header';
 describe('Header', () => {
   test('match snapshots', () => {
     const renderer = new ShallowRenderer();
-    renderer.render(<Header appBaseUrl="http://foo.de" />);
+    renderer.render(<Header />);
     const tree = renderer.getRenderOutput();
     expect(tree).toMatchSnapshot();
     // Login & LanguageSelect are snapshoted as 'UNDEFINED' due to React.Memo