+++ src/layers/ZweitausbildungRoutesHighlightLayer/ZweitausbildungRoutesHighlightLayer.js
@@ -3,8 +3,8 @@ import { MapboxStyleLayer } from 'mobility-toolbox-js/ol';
 import IconList from '../../components/IconList';
 import lines from '../ZweitausbildungRoutesLayer/lines';
 
-const sourceId = 'base';
-const sourceLayer = 'osm_edges';
+const sourceId = 'ch.sbb.zweitausbildung';
+const sourceLayer = 'ch.sbb.zweitausbildung';
 
 /**
  * Layer for zweitausbildung highlight routes.
@@ -153,7 +153,6 @@ class ZweitausbildungRoutesHighlightLayer extends MapboxStyleLayer {
     if (!mbMap) {
       return;
     }
-
     this.styleLayers.forEach(({ id }) => {
       mbMap.setPaintProperty(id, 'line-color', 'rgba(0,0,0,0)');
     });
@@ -172,13 +171,15 @@ class ZweitausbildungRoutesHighlightLayer extends MapboxStyleLayer {
       return;
     }
 
-    const { color } = lines[this.selected];
+    const { color } = lines[this.selected] || {};
     if (!color) {
-      // eslint-disable-next-line no-console
-      console.log(
-        `There is no color defined for ${this.selected}, available labels are `,
-        lines,
-      );
+      if (this.selected !== 'Alle') {
+        // eslint-disable-next-line no-console
+        console.log(
+          `There is no color defined for ${this.selected}, available labels are `,
+          lines,
+        );
+      }
       return;
     }